requirements.txt
neo4j>=5.12.0,<6
openai>=1.0.0,<2
python-dotenv>=1.0.0,<2
